+++ source/global.d.ts
@@ -77,6 +77,10 @@ declare const PROJECT_PROPERTIES_WEBPACK_ENTRY: string
 declare const SPLASH_SCREEN_WEBPACK_ENTRY: string
 declare const SPLASH_SCREEN_PRELOAD_WEBPACK_ENTRY: string
 
+// Contains the git build number
+// eslint-disable-next-line @typescript-eslint/naming-convention
+declare const __GIT_COMMIT_HASH__: string
+
 /**
  * Declare and extend the global NodeJS object to enable the globals
  * for the service providers.